+++ src/_pytask/live.py
@@ -94,12 +94,22 @@ class LiveManager:
     """A class for live displays during a session.
 
     This class allows to display live information during a session and handles the
-    interaction with the :class:`_pytask.debugging.PytaskPDB` and
-    :class:`_pytask.capture.CaptureManager`.
+    interaction with the :class:`_pytask.debugging.PytaskPDB`.
+
+    The renderable is not updated automatically for two reasons.
+
+    1. Usually, the duration of tasks is highly heterogeneous and there are probably not
+       many tasks which last much less than a second. Therefore, updating the renderable
+       automatically by a fixed time interval seems unnecessary.
+
+    2. To update the renderable automatically a thread is started which pushes the
+       updates. When a task is run simultaneously and capturing is activated, all
+       updates will be captured and added to the stdout of the task instead of printed
+       to the terminal.
 
     """
 
-    _live = Live(renderable=None, console=console, auto_refresh=True)
+    _live = Live(renderable=None, console=console, auto_refresh=False)
 
     def start(self) -> None:
         self._live.start()
@@ -121,6 +131,7 @@ class LiveManager:
 
     def update(self, *args: Any, **kwargs: Any) -> None:
         self._live.update(*args, **kwargs)
+        self._live.refresh()
 
     @property
     def is_started(self) -> None:
@@ -144,8 +155,10 @@ class LiveExecution:
         end."""
         self._live_manager.start()
         yield
-        self._update_table(reduce_table=False, sort_table=True)
-        self._live_manager.stop(transient=False)
+        self._live_manager.stop(transient=True)
+        table = self._generate_table(reduce_table=False, sort_table=True)
+        if table is not None:
+            console.print(table)
 
     @hookimpl(tryfirst=True)
     def pytask_execute_task_log_start(self, task: MetaTask) -> bool:
@@ -168,53 +181,53 @@ class LiveExecution:
         if more entries are requested, the list is filled up with completed tasks.
 
         """
-        if self._running_tasks or self._reports:
+        n_reports_to_display = self._n_entries_in_table - len(self._running_tasks)
+        if not reduce_table:
+            relevant_reports = self._reports
+        elif n_reports_to_display >= 1:
+            relevant_reports = self._reports[-n_reports_to_display:]
+        else:
+            relevant_reports = []
 
-            n_reports_to_display = self._n_entries_in_table - len(self._running_tasks)
-            if not reduce_table:
-                relevant_reports = self._reports
-            elif n_reports_to_display >= 1:
-                relevant_reports = self._reports[-n_reports_to_display:]
-            else:
-                relevant_reports = []
+        if sort_table:
+            relevant_reports = sorted(
+                relevant_reports, key=lambda report: report["name"]
+            )
 
-            if sort_table:
-                relevant_reports = sorted(
-                    relevant_reports, key=lambda report: report["name"]
+        table = Table()
+        table.add_column("Task", overflow="fold")
+        table.add_column("Outcome")
+        for report in relevant_reports:
+            if (
+                report["outcome"]
+                in (
+                    TaskOutcome.SKIP,
+                    TaskOutcome.SKIP_UNCHANGED,
+                    TaskOutcome.SKIP_PREVIOUS_FAILED,
+                    TaskOutcome.PERSISTENCE,
                 )
-
-            table = Table()
-            table.add_column("Task", overflow="fold")
-            table.add_column("Outcome")
-            for report in relevant_reports:
-                if (
-                    report["outcome"]
-                    in (
-                        TaskOutcome.SKIP,
-                        TaskOutcome.SKIP_UNCHANGED,
-                        TaskOutcome.SKIP_PREVIOUS_FAILED,
-                        TaskOutcome.PERSISTENCE,
-                    )
-                    and self._verbose < 2
-                ):
-                    pass
-                else:
-                    table.add_row(
-                        format_task_id(
-                            report["task"],
-                            editor_url_scheme=self._editor_url_scheme,
-                            short_name=True,
-                        ),
-                        Text(report["outcome"].symbol, style=report["outcome"].style),
-                    )
-            for task in self._running_tasks.values():
+                and self._verbose < 2
+            ):
+                pass
+            else:
                 table.add_row(
                     format_task_id(
-                        task, editor_url_scheme=self._editor_url_scheme, short_name=True
+                        report["task"],
+                        editor_url_scheme=self._editor_url_scheme,
+                        short_name=True,
                     ),
-                    "running",
+                    Text(report["outcome"].symbol, style=report["outcome"].style),
                 )
-        else:
+        for task in self._running_tasks.values():
+            table.add_row(
+                format_task_id(
+                    task, editor_url_scheme=self._editor_url_scheme, short_name=True
+                ),
+                "running",
+            )
+
+        # If the table is empty, do not display anything.
+        if table.rows == []:
             table = None
 
         return table
@@ -254,21 +267,24 @@ class LiveCollection:
 
     @hookimpl(hookwrapper=True)
     def pytask_collect(self) -> Generator[None, None, None]:
+        """Start the status of the cllection."""
         self._live_manager.start()
         yield
 
     @hookimpl
     def pytask_collect_file_log(self, reports: List[CollectionReport]) -> None:
+        """Update the status after a file is collected."""
         self._update_statistics(reports)
         self._update_status()
 
     @hookimpl(hookwrapper=True)
     def pytask_collect_log(self) -> Generator[None, None, None]:
-        self._live_manager.update(None)
+        """Stop the live display when all tasks have been collected."""
         self._live_manager.stop(transient=True)
         yield
 
     def _update_statistics(self, reports: List[CollectionReport]) -> None:
+        """Update the statistics on collected tasks and errors."""
         if reports is None:
             reports = []
         for report in reports:
@@ -278,10 +294,12 @@ class LiveCollection:
                 self._n_errors += 1
 
     def _update_status(self) -> None:
+        """Update the status."""
         status = self._generate_status()
         self._live_manager.update(status)
 
     def _generate_status(self) -> Status:
+        """Generate the status."""
         msg = f"Collected {self._n_collected_tasks} tasks."
         if self._n_errors > 0:
             msg += f" {self._n_errors} errors."