+++ tests/test_clean.py
@@ -1,6 +1,7 @@
 import textwrap
 
 import pytest
+from _pytask.outcomes import ExitCode
 from pytask import cli
 
 
@@ -153,7 +154,7 @@ def test_configuration_failed(runner, tmp_path):
     result = runner.invoke(
         cli, ["clean", tmp_path.joinpath("non_existent_path").as_posix()]
     )
-    assert result.exit_code == 2
+    assert result.exit_code == ExitCode.CONFIGURATION_FAILED
 
 
 @pytest.mark.end_to_end
@@ -164,4 +165,4 @@ def test_collection_failed(runner, tmp_path):
     tmp_path.joinpath("task_module.py").write_text(textwrap.dedent(source))
 
     result = runner.invoke(cli, ["clean", tmp_path.as_posix()])
-    assert result.exit_code == 3
+    assert result.exit_code == ExitCode.COLLECTION_FAILED