+++ tests/test_console.py
@@ -1,8 +1,9 @@
 import inspect
 from pathlib import Path
 
-import attr
 import pytest
+from _pytask.console import _get_file
+from _pytask.console import _get_source_lines
 from _pytask.console import console
 from _pytask.console import create_summary_panel
 from _pytask.console import create_url_style_for_path
@@ -10,7 +11,6 @@ from _pytask.console import create_url_style_for_task
 from _pytask.console import format_task_id
 from _pytask.console import render_to_string
 from _pytask.nodes import create_task_name
-from _pytask.nodes import MetaTask
 from _pytask.nodes import PythonFunctionTask
 from _pytask.outcomes import CollectionOutcome
 from _pytask.outcomes import TaskOutcome
@@ -19,23 +19,11 @@ from rich.style import Style
 from rich.text import Span
 from rich.text import Text
 
+from tests._test_console_helpers import empty_decorator
+
 
 def task_func():
     ...
-
-
-@attr.s
-class DummyTask(MetaTask):
-    function = attr.ib()
-
-    def state():
-        ...
-
-    def execute():
-        ...
-
-    def add_report_section():
-        ...
 
 
 _SOURCE_LINE_TASK_FUNC = inspect.getsourcelines(task_func)[1]
@@ -57,8 +45,7 @@ _SOURCE_LINE_TASK_FUNC = inspect.getsourcelines(task_func)[1]
 )
 def test_create_url_style_for_task(edtior_url_scheme, expected):
     path = Path(__file__)
-    task = DummyTask(task_func)
-    style = create_url_style_for_task(task, edtior_url_scheme)
+    style = create_url_style_for_task(task_func, edtior_url_scheme)
     assert style == Style.parse(expected.format(path=path))
 
 
@@ -104,15 +91,16 @@ def test_create_summary_panel(capsys, outcome, outcome_enum, total_description):
 
 @pytest.mark.unit
 @pytest.mark.parametrize(
-    "color_system, text, expected",
+    "color_system, text, strip_styles, expected",
     [
-        (None, "[red]text", "text\n"),
-        ("truecolor", "[red]text", "\x1b[31mtext\x1b[0m\n"),
+        (None, "[red]text", False, "text\n"),
+        ("truecolor", "[red]text", False, "\x1b[31mtext\x1b[0m\n"),
+        ("truecolor", "[red]text", True, "text\n"),
     ],
 )
-def test_render_to_string(color_system, text, expected):
+def test_render_to_string(color_system, text, strip_styles, expected):
     console = Console(color_system=color_system)
-    result = render_to_string(text, console)
+    result = render_to_string(text, console=console, strip_styles=strip_styles)
     assert result == expected
 
 
@@ -184,4 +172,37 @@ def test_format_task_id(
         task.short_name = short_name
 
     result = format_task_id(task, editor_url_scheme, use_short_name, relative_to)
+    assert result == expected
+
+
+@pytest.mark.parametrize(
+    "task_func, skipped_paths, expected",
+    [
+        (task_func, [], _THIS_FILE),
+        (
+            empty_decorator(task_func),
+            [],
+            _THIS_FILE.parent.joinpath("_test_console_helpers.py"),
+        ),
+        (
+            empty_decorator(task_func),
+            [_THIS_FILE.parent.joinpath("_test_console_helpers.py")],
+            _THIS_FILE,
+        ),
+    ],
+)
+def test_get_file(task_func, skipped_paths, expected):
+    result = _get_file(task_func, skipped_paths)
+    assert result == expected
+
+
+@pytest.mark.parametrize(
+    "task_func, expected",
+    [
+        (task_func, _SOURCE_LINE_TASK_FUNC),
+        (empty_decorator(task_func), _SOURCE_LINE_TASK_FUNC),
+    ],
+)
+def test_get_source_lines(task_func, expected):
+    result = _get_source_lines(task_func)
     assert result == expected